Deleted unnecessary classes and moved objects to their proper modules.
