+++ app/src/main/java/com/teamtwo/apilol/model/LOLService.kt
@@ -2,7 +2,7 @@ package com.teamtwo.apilol.model
 
 import com.example.domain.ChampionsResponse
 import com.teamtwo.apilol.model.spells.SpellsResponseModel
-import com.teamtwo.apilol.model.matches.FeaturedGamesResponse
+import com.example.domain.FeaturedGamesResponse
 import com.teamtwo.apilol.model.items.ItemsResponse
 import com.teamtwo.apilol.model.sumonners.Summoner
 import retrofit2.Response
