+++ domain/src/main/java/com/example/domain/MatchesModels.kt
@@ -1,4 +1,4 @@
-package com.teamtwo.apilol.model.matches
+package com.example.domain
 
 data class FeaturedGamesResponse (
     val clientRefreshInterval: Long?,
@@ -37,5 +37,46 @@ data class ParticipantService (
     val spell2Id: Long?,
     val teamId: Long?,
     val spell1Id: Long?
+)
+
+data class FeaturedGameInfo (
+    val gameId: Long,
+    val gameStartTime: Long,
+    val platformId: String,
+    val gameMode: String,
+    val mapId: Long,
+    val gameType: String,
+    val bannedChampions: List<BannedChampion>,
+    val observers: Observer,
+    val participants: List<Participant>,
+    val gameLength: Long,
+    val gameQueue: Queue?
+)
+
+data class BannedChampion (
+    val pickTurn: Int,
+    val championId: Long,
+    val teamId: Long
+)
+
+data class Observer (
+    val encryptionKey: String
+)
+
+data class Participant (
+    val profileIconId: Long,
+    val championId: Long,
+    val summonerName: String,
+    val bot: Boolean,
+    val spell2Id: Long,
+    val teamId: Long,
+    val spell1Id: Long
+)
+
+data class Queue (
+    val queueId: Int,
+    val map: String,
+    val description: String?,
+    val notes: String?
 )
